requirements.txt
python-dotenv
spotipy
openai
